test: attempt at resolving qlty check issues
